% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{allele_frequencies}
\alias{allele_frequencies}
\title{Reference-allele frequencies of a genotype matrix}
\usage{
allele_frequencies(X)
}
\arguments{
\item{X}{Genotype matrix, individuals in rows and variants in columns, coded
as the number of copies of the reference allele.}
}
\value{
A numeric vector of allele frequencies, one per column.
}
\description{
Reference-allele frequencies of a genotype matrix
}
