% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theory.R
\name{split_effect_sizes}
\alias{split_effect_sizes}
\title{Common and rare effect sizes for a given rare-to-common ratio}
\usage{
split_effect_sizes(h2, tau2, p_common, p_rare, s)
}
\arguments{
\item{h2}{The target heritability, in \verb{(0, 1)}.}

\item{tau2}{The non-genetic variance.}

\item{p_common}{The reference-allele frequencies of the common causal
variants.}

\item{p_rare}{The reference-allele frequencies of the rare causal variants.}

\item{s}{The rare-to-common effect-size ratio.}
}
\value{
A list with elements \code{common} and \code{rare}, the two effect sizes.
}
\description{
Rare causal variants are given an effect \code{s} times that of common ones.
Writing \code{beta_c} for the common effect size, the additive genetic variance is
\verb{2 beta_c^2 sum_c p_c (1 - p_c) + 2 s^2 beta_c^2 sum_r p_r (1 - p_r)}, so
matching a target heritability \code{h2} at non-genetic variance \code{tau^2} gives
}
\details{
\verb{beta_c = sqrt( h2/(1-h2) * tau^2 / ( 2 sum_c p_c(1-p_c) + 2 s^2 sum_r p_r(1-p_r) ) )}.
}
\examples{
p_common <- runif(500, 0.05, 0.95)
p_rare <- runif(500, 0.001, 0.01)
# rare variants act five times as strongly as common ones
split_effect_sizes(h2 = 0.5, tau2 = 1, p_common, p_rare, s = 5)

}
